[refactor] BarcodeScanner 컴포넌트 구조 개선 및 커스텀 훅 분리
- BarcodeScanner를 ScannerHeader, ScannerControls 컴포넌트로 분리
- 바코드 스캔 로직을 useBarcodeScanner 훅으로 추출
- 위치 추적 로직을 usePositionTracking 훅으로 분리
- 스캐너 설정 상수를 별도 파일로 분리 (SCANNER_CONFIG)
- TTS 관련 로직을 전역 스토어(useTTSStore)로 통합
- FacingMode 타입 정의 추가 및 타입 안정성 개선
- CertSection 코드 스타일 개선
