+++ src/app/scan/_hooks/useVibrate.ts
@@ -1,0 +1,11 @@
+import { useCallback } from 'react'
+
+export function useVibrate() {
+  const vibrate = useCallback((pattern: number | number[]) => {
+    if (navigator.vibrate) {
+      navigator.vibrate(pattern)
+    }
+  }, [])
+
+  return { vibrate }
+}
